+++ src/main/java/net/geforcemods/securitycraft/ClientHandler.java
@@ -34,6 +34,7 @@ import net.geforcemods.securitycraft.renderers.SecurityCameraRenderer;
 import net.geforcemods.securitycraft.renderers.SentryRenderer;
 import net.geforcemods.securitycraft.renderers.SonicSecuritySystemRenderer;
 import net.geforcemods.securitycraft.renderers.TrophySystemRenderer;
+import net.geforcemods.securitycraft.screen.BlockChangeDetectorScreen;
 import net.geforcemods.securitycraft.screen.BlockPocketManagerScreen;
 import net.geforcemods.securitycraft.screen.BlockReinforcerScreen;
 import net.geforcemods.securitycraft.screen.BriefcaseInventoryScreen;
@@ -117,6 +118,7 @@ public class ClientHandler {
 		};
 		//@formatter:off
 		ResourceLocation[] facingPoweredBlocks = {
+				new ResourceLocation(SecurityCraft.MODID, "block_change_detector"),
 				new ResourceLocation(SecurityCraft.MODID, "keycard_reader"),
 				new ResourceLocation(SecurityCraft.MODID, "keypad"),
 				new ResourceLocation(SecurityCraft.MODID, "retinal_scanner")
@@ -242,6 +244,7 @@ public class ClientHandler {
 		RenderType cutoutMipped = RenderType.cutoutMipped();
 		RenderType translucent = RenderType.translucent();
 
+		ItemBlockRenderTypes.setRenderLayer(SCContent.BLOCK_CHANGE_DETECTOR.get(), cutout);
 		ItemBlockRenderTypes.setRenderLayer(SCContent.BLOCK_POCKET_MANAGER.get(), cutoutMipped);
 		ItemBlockRenderTypes.setRenderLayer(SCContent.BLOCK_POCKET_WALL.get(), translucent);
 		ItemBlockRenderTypes.setRenderLayer(SCContent.CAGE_TRAP.get(), cutoutMipped);
@@ -330,6 +333,7 @@ public class ClientHandler {
 			MenuScreens.register(SCContent.mTypeBlockPocketManager, BlockPocketManagerScreen::new);
 			MenuScreens.register(SCContent.mTypeProjector, ProjectorScreen::new);
 			MenuScreens.register(SCContent.mTypeTrophySystem, TrophySystemScreen::new);
+			MenuScreens.register(SCContent.mTypeBlockChangeDetector, BlockChangeDetectorScreen::new);
 		});
 		KeyBindings.init();
 		cameraOverlay = OverlayRegistry.registerOverlayTop(SecurityCraft.MODID + ":camera_overlay", SCClientEventHandler::cameraOverlay);
@@ -356,6 +360,7 @@ public class ClientHandler {
 		event.registerBlockEntityRenderer(SCContent.beTypeSonicSecuritySystem, SonicSecuritySystemRenderer::new);
 		event.registerBlockEntityRenderer(SCContent.beTypeTrophySystem, TrophySystemRenderer::new);
 		//disguisable block entity renderers
+		event.registerBlockEntityRenderer(SCContent.beTypeBlockChangeDetector, DisguisableBlockEntityRenderer::new);
 		event.registerBlockEntityRenderer(SCContent.beTypeCageTrap, DisguisableBlockEntityRenderer::new);
 		event.registerBlockEntityRenderer(SCContent.beTypeInventoryScanner, DisguisableBlockEntityRenderer::new);
 		event.registerBlockEntityRenderer(SCContent.beTypeKeycardReader, DisguisableBlockEntityRenderer::new);
@@ -463,7 +468,8 @@ public class ClientHandler {
 
 			return noTint;
 			//@formatter:off
-		}, SCContent.CAGE_TRAP.get(),
+		}, SCContent.BLOCK_CHANGE_DETECTOR.get(),
+				SCContent.CAGE_TRAP.get(),
 				SCContent.INVENTORY_SCANNER.get(),
 				SCContent.KEYCARD_READER.get(),
 				SCContent.KEYPAD.get(),