+++ src/main/java/net/geforcemods/securitycraft/blockentities/ProjectorBlockEntity.java
@@ -202,9 +202,10 @@ public class ProjectorBlockEntity extends DisguisableBlockEntity implements Cont
 		if (count >= 1) {
 			projectedBlock = ItemStack.EMPTY;
 			resetSavedState();
+			return stack;
 		}
 
-		return stack;
+		return ItemStack.EMPTY;
 	}
 
 	@Override