+++ src/main/java/net/geforcemods/securitycraft/blockentities/SonicSecuritySystemBlockEntity.java
@@ -12,9 +12,9 @@ import net.geforcemods.securitycraft.api.ILockable;
 import net.geforcemods.securitycraft.api.Option;
 import net.geforcemods.securitycraft.api.Option.IntOption;
 import net.geforcemods.securitycraft.blocks.SonicSecuritySystemBlock;
+import net.geforcemods.securitycraft.misc.BlockEntityTracker;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.misc.SCSounds;
-import net.geforcemods.securitycraft.misc.SonicSecuritySystemTracker;
 import net.geforcemods.securitycraft.util.BlockUtils;
 import net.geforcemods.securitycraft.util.ITickingBlockEntity;
 import net.minecraft.core.BlockPos;
@@ -74,7 +74,7 @@ public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity impl
 	public void tick(Level level, BlockPos pos, BlockState state) {
 		// Add this SSS to the global tracker if it has not already been added
 		if (!tracked) {
-			SonicSecuritySystemTracker.track(this);
+			BlockEntityTracker.SONIC_SECURITY_SYSTEM.track(this);
 			tracked = true;
 		}
 
@@ -150,7 +150,7 @@ public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity impl
 		super.setRemoved();
 
 		// Stop tracking SSSs when they are removed from the world
-		SonicSecuritySystemTracker.stopTracking(this);
+		BlockEntityTracker.SONIC_SECURITY_SYSTEM.stopTracking(this);
 	}
 
 	@Override